requirements.txt
flask==3.0.2
werkzeug==3.0.1
python-dotenv==1.0.1
waitress==3.0.0
flask-limiter==3.5.1
flask-cors>=4.0.0
Pillow==10.3.0
python-dateutil==2.9.0
anthropic>=0.40.0
boto3>=1.34.0
